chore(core): reduce internal exports
