+++ packages/core/src/provider/createContext.ts
@@ -11,7 +11,7 @@ import {
 } from '@rsbuild/shared';
 import type { InternalContext } from '../types';
 import { initHooks } from '../initHooks';
-import { withDefaultConfig } from './config';
+import { withDefaultConfig } from '../config';
 import { getEntryObject } from '../plugins/entry';
 
 function getAbsolutePath(root: string, filepath: string) {
