--- packages/core/src/provider/index.ts
@@ -1,14 +1,0 @@
-export { rspackProvider } from './provider';
-export type { Rspack, RspackConfig } from '@rsbuild/shared';
-export { createContext, createPublicContext } from './createContext';
-export { initPlugins, createPluginManager } from '../pluginManager';
-export { initHooks, type Hooks } from '../initHooks';
-export { initRsbuildConfig } from './initConfigs';
-export { getPluginAPI } from './initPlugins';
-export { applyBaseCSSRule, applyCSSModuleRule } from './plugins/css';
-export type { InternalContext } from '../types';
-export { setHTMLPlugin, getHTMLPlugin } from './htmlPluginUtil';
-export { formatStats, getStatsOptions } from './shared';
-export { getChainUtils } from './rspackConfig';
-export { applySwcDecoratorConfig } from './plugins/swc';
-export { getDevMiddleware } from './devMiddleware';
