+++ packages/core/src/provider/initConfigs.ts
@@ -10,7 +10,7 @@ import { mergeRsbuildConfig } from '../mergeConfig';
 import { updateContextByNormalizedConfig } from './createContext';
 import { inspectConfig } from './inspectConfig';
 import { generateRspackConfig } from './rspackConfig';
-import { normalizeConfig } from './config';
+import { normalizeConfig } from '../config';
 import { initPlugins } from '../pluginManager';
 import type { InternalContext, NormalizedConfig } from '../types';
