+++ packages/core/src/rspack/InlineChunkHtmlPlugin.ts
@@ -14,7 +14,7 @@ import {
 } from '@rsbuild/shared';
 import type { Compiler, Compilation } from '@rspack/core';
 import type { HtmlTagObject } from 'html-webpack-plugin';
-import { getHTMLPlugin } from '../provider/htmlPluginUtil';
+import { getHTMLPlugin } from '../htmlUtils';
 
 export type InlineChunkHtmlPluginOptions = {
   styleTests: InlineChunkTest[];
